+++ Source/GameLoop/Game.cpp
@@ -48,17 +48,8 @@ void Game::Start() {
 	ui->CrateTempFontFromTTF("Fonts/arial.ttf", 20, "arial20");
 	ui->CrateTempFontFromTTF("Fonts/arial.ttf", 40, "arial40");
 
-	MT::Texture* tex = TexMan::GetTex("arial12");
-	MT::Texture* tex2 = TexMan::GetTex("arial20");
-	MT::Texture* tex3 = TexMan::GetTex("arial40");
-
-	std::println("Arial12 Width: {} Heigth: {}", tex->w, tex->h);
-	std::println("Arial20 Width: {} Heigth: {}", tex2->w, tex2->h);
-	std::println("Arial40 Width: {} Heigth: {}", tex3->w, tex3->h);
-
-	Button *btn = ui->CreateButtonF("btn1", 100, 100, 100, 100, nullptr,"arial12","Some text + = 1");
-	btn->SetColor(30, 30, 30);
-	btn->SetRenderTextType((int)TextRenderType::Centered);
+	FileExplorer fe;
+	fe.Open();
 
 }
 
@@ -82,12 +73,6 @@ void Game::Input() {
 
 void Game::Render() {
 	renderer->ClearFrame(Global::defaultDrawColor[0], Global::defaultDrawColor[1], Global::defaultDrawColor[2]);
-	MT::Texture* tex = TexMan::GetTex("arial12");
-	MT::Texture* tex2 = TexMan::GetTex("arial20");
-	MT::Texture* tex3 = TexMan::GetTex("arial40");
-	renderer->RenderCopyFiltered({ 10,10,(int)tex->w,(int)tex->h }, TexMan::GetTex("arial12"), {255,0,0});
-	renderer->RenderCopyFiltered({ 10,200,(int)tex2->w,(int)tex2->h }, TexMan::GetTex("arial20"), { 255,0,0 });
-	renderer->RenderCopyFiltered({ 10,400,(int)tex3->w,(int)tex3->h }, TexMan::GetTex("arial40"), { 255,0,0 });
 	ui->Render();
 	renderer->Present();
 }