+++ Source/GameLoop/main.cpp
@@ -1,8 +1,7 @@
 #include "Game.h"
 #include "GlobalVariables.h"
 
-int main(int argv, char* argc[])
-{
+int main(int argv, char* argc[]){
 	//srand(time(0));
 	srand(0);
 	Game game;